requirements.txt
discord.py>=0.16.12
emoji==0.5.4
python-dotenv==0.10.3
requests==2.22.0
APScheduler==3.6.3
